Add dynamically rendered tabs and tab controls
